refact some code
