+++ src/webportal/src/app/home/home/conn.js
@@ -18,8 +18,6 @@
 import querystring from 'querystring';
 
 import config from '../../config/webportal.config';
-import {isNil} from 'lodash';
-import {isLongRunJob} from '../../components/util/job';
 
 const username = cookies.get('user');
 const token = cookies.get('token');
@@ -47,6 +45,10 @@ async function fetchWrapper(...args) {
 
 export async function listJobs() {
   return fetchWrapper(`${config.restServerUri}/api/v1/jobs?${querystring.stringify({username})}`);
+}
+
+export async function listAllJobs() {
+  return fetchWrapper(`${config.restServerUri}/api/v1/jobs`);
 }
 
 export async function getUserInfo() {
@@ -81,21 +83,15 @@ export async function getAvailableGpuPerNode() {
   }
 }
 
-export async function listAbnormalJobs() {
+export async function getLowGpuJobInfos() {
   const prometheusQuery = `avg(avg_over_time(task_gpu_percent[10m]) < 10) by (job_name)`;
-  const results = await Promise.all([
-    fetchWrapper(`${config.restServerUri}/api/v1/jobs`),
-    fetch(`${config.prometheusUri}/api/v1/query?query=${encodeURIComponent(prometheusQuery)}`),
-  ]);
-  const allRuuingJobs = results[0].filter((job) => job.state === 'RUNNING');
-  const longRunJobs = allRuuingJobs.filter(isLongRunJob);
-
-  if (!results[1].ok) {
-    return longRunJobs;
+  const res = await fetch(`${config.prometheusUri}/api/v1/query?query=${encodeURIComponent(prometheusQuery)}`);
+  const json = await res.json();
+  if (!res.ok) {
+    throw new Error(json.message);
   }
 
-  const json = await results[1].json();
-  const gpuUsageInfos = json.data.result.map((keyValuePair) => {
+  const lowGpuJobInfos = json.data.result.map((keyValuePair) => {
     const frameworkName = keyValuePair.metric.job_name;
     const jobNameBeginIndex = frameworkName.indexOf('~');
     return {
@@ -103,32 +99,6 @@ export async function listAbnormalJobs() {
       gpuUsage: keyValuePair.value[1],
     };
   });
-
-  // Get low GPU usage jobs
-  const lowGpuUsageJobs = allRuuingJobs.reduce((acc, cur)=>{
-    const gpuUsageInfo = gpuUsageInfos.find((info) => info.jobName === cur.name);
-    if (isNil(gpuUsageInfo)) {
-      return acc;
-    }
-    const lowGpuUsageJob = {...cur};
-    lowGpuUsageJob['gpuUsage'] = gpuUsageInfo.gpuUsage;
-    acc.push(lowGpuUsageJob);
-    return acc;
-  }, []);
-
-  // Merge long run jobs and low GPU usage jobs
-  const abnormalJobs = [...longRunJobs];
-  abnormalJobs.forEach((job) => {
-    const lowGpuUsagejob = lowGpuUsageJobs.find((lowGpuUsageJob) => lowGpuUsageJob.name === job.name);
-    if (!isNil(lowGpuUsagejob)) {
-      job['gpuUsage'] = lowGpuUsagejob.gpuUsage;
-    }
-  });
-  lowGpuUsageJobs.forEach((lowGpuUsageJob) => {
-    if (isNil(abnormalJobs.find((job) => job.name === lowGpuUsageJob.name))) {
-      abnormalJobs.push(lowGpuUsageJob);
-    }
-  });
-  return abnormalJobs;
+  return lowGpuJobInfos;
 }
